app/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
mne>=1.10.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow-macos>=2.8.0
tensorflow-metal>=0.5.0
seaborn>=0.11.0
flask>=2.0.0
joblib>=1.1.0
PyWavelets>=1.4.0
streamlit>=1.10.0